removeing release_mem_region because common/driver src code will be updated to devm_request_mem_region
When you use devm_request_mem_region to request a memory region, the memory is managed automatically by the device management (devm) API within the Linux kernel. This means that the memory region will be automatically released when the device is detached or the driver is unloaded. Therefore, you do not need to explicitly call a function to release the memory region when using devm_request_mem_region.