Improve error handling in DataDev_Probe and DataGpu_Probe
Kernel docs say if dma_set_mask or dma_set_coherent_mask fail, the driver
should avoid using DMA at all for the device. That is now considered an
error in this code.

Unified the error handling paths to reduce duplicated code

On error, we also now clear the slot we used in gDmaDevices. This avoids
a super rare failure case where we run out of slots in the array due to
cards that fail to init properly (i.e. dma_set_xxx failing, or some other
error path in DataXXX_Probe)