Disable IRQ after request_irq call, fix irqDisable flag
request_irq does not disable the IRQ by default. It is possible to get
into a situation where an IRQ arrives and begins processing before the
IRQ handler has been completely setup. This leads to a unrecoverable
page fault in the kernel.

Additionally, the irqDisable flag was accidentally broken at some point.