Fix coherent dma on 32bit ARM for newer kernel versions
Related commit in the kernel seems to be ae626eb97376148bb63c3f3ca9517fde0f39bec3.
Apparently included in V6.0 and later, so a kernel version check is
introduced. For newer kernel versions, coherent dma may be enabled by
adding the dma-coherent attribute to the dma device in the device tree.
For the case of coherent dma being set in the device tree (v6.0 and
upwards) an info message was added to confirm whether coherent dma is
enabled or not.